raise InvalidFunctionError on invalid idl
Raise a custom error type when a function definition is invalid instead
of simply dying on a key error. The new error types are consumed in
stackhut_toolkit to display a helpful error to the user and to prevent
the internal error handling (stacktrace, apology, call to action, and
logging the exception remotely).

- prevent Function constructor from crashing when accessing the 'returns'
key for a function dict that does not have one (happens if the function
definition omits the return type).
- add validation to function constructor that raises
InvalidFunctionError if critical information is missing (currently only
if the name is blank or None or if the return type was not parsed).

- add barrister.exceptions with ConfigError and InvalidFunctionError
- expose barrister.exceptions as stackhut_common.exceptions